REFACTOR - Pushed OAuth to a seperate ClassLibrary

+++ Twitch.Api/Model/StreamElements/StreamStream.cs
@@ -9,7 +9,7 @@ namespace Twitch.Api.Model.StreamElements
         public float Average_Fps { set; get; }
         public int Video_Height { set; get; }
         public DateTime Created_At { set; get; }
-        public int _Id { set; get; }
+        public long _Id { set; get; }
         public Channel Channel { set; get; }
         public StreamStreamLinks _Links { set; get; }
     }
